move contract wrapper into their own subfolder
