+++ lib/arc.js
@@ -1,10 +1,10 @@
-export * from "./absoluteVote.js";
+export * from "./contracts/absoluteVote.js";
 export * from "./config.js";
 export * from "./contracts.js";
-export * from "./contributionreward.js";
-export * from "./globalconstraintregistrar.js";
+export * from "./contracts/contributionreward.js";
+export * from "./contracts/globalconstraintregistrar.js";
 export * from "./dao.js";
-export * from "./schemeregistrar.js";
-export * from "./tokenCapGC.js";
-export * from "./upgradescheme.js";
+export * from "./contracts/schemeregistrar.js";
+export * from "./contracts/tokenCapGC.js";
+export * from "./contracts/upgradescheme.js";
 export * from "./utils.js";
