+++ lib/contracts.js
@@ -4,13 +4,13 @@ import { requireContract } from "./utils.js";
  * These are uninitialized instances of ExtendTruffleContract,
  * effectively class factories.
  */
-import { AbsoluteVote } from "./absoluteVote.js";
-import { ContributionReward } from "./contributionreward.js";
+import { AbsoluteVote } from "./contracts/absoluteVote.js";
+import { ContributionReward } from "./contracts/contributionreward.js";
 const GenesisScheme = requireContract("GenesisScheme");
-import { GlobalConstraintRegistrar } from "./globalconstraintregistrar.js";
-import { SchemeRegistrar } from "./schemeregistrar.js";
-import { TokenCapGC } from "./tokenCapGC.js";
-import { UpgradeScheme } from "./upgradescheme.js";
+import { GlobalConstraintRegistrar } from "./contracts/globalconstraintregistrar.js";
+import { SchemeRegistrar } from "./contracts/schemeregistrar.js";
+import { TokenCapGC } from "./contracts/tokenCapGC.js";
+import { UpgradeScheme } from "./contracts/upgradescheme.js";
 
 export async function getDeployedContracts() {
   /**
