+++ lib/contracts/contributionreward.js
@@ -6,7 +6,7 @@ import {
   ExtendTruffleContract,
   requireContract,
   getWeb3
-} from "./utils.js";
+} from "../utils.js";
 
 const SolidityContributionReward = requireContract("ContributionReward");
