+++ lib/contracts/globalconstraintregistrar.js
@@ -4,7 +4,7 @@ const dopts = require("default-options");
 import {
   ExtendTruffleContract,
   requireContract
-} from "./utils.js";
+} from "../utils.js";
 
 const SolidityGlobalConstraintRegistrar = requireContract(
   "GlobalConstraintRegistrar"
