+++ lib/contracts/schemeregistrar.js
@@ -1,11 +1,11 @@
 "use strict";
 const dopts = require("default-options");
 
-import { getDeployedContracts } from "./contracts.js";
+import { getDeployedContracts } from "../contracts.js";
 import {
   ExtendTruffleContract,
   requireContract
-} from "./utils.js";
+} from "../utils.js";
 
 const SoliditySchemeRegistrar = requireContract("SchemeRegistrar");
