+++ lib/contracts/tokenCapGC.js
@@ -1,7 +1,7 @@
 "use strict";
 // const dopts = require('default-options');
 
-import { ExtendTruffleContract, requireContract } from "./utils.js";
+import { ExtendTruffleContract, requireContract } from "../utils.js";
 
 const SolidityTokenCapGC = requireContract("TokenCapGC");
